the global ID of reactves and keychains is now a long
yay for more than 4 billion turns!
benchmarks say that we can have about 1 million turns per second
so we could have exhausted the number of turns in just an hour or so
but now we are closer to half a million … years

+++ REScala/src/rescala/graph/Data.scala
@@ -1,6 +1,6 @@
 package rescala.graph
 
-import java.util.concurrent.atomic.AtomicInteger
+import java.util.concurrent.atomic.AtomicLong
 
 import scala.util.DynamicVariable
 
@@ -27,7 +27,7 @@ object Globals {
   val dynamicNameVar = new DynamicVariable("")
   def named[S](n: String)(f: => S): S = dynamicNameVar.withValue(n)(f)
 
-  private val counter = new AtomicInteger(0)
+  private val counter = new AtomicLong(0)
   def nextID() = counter.incrementAndGet()
 }
